# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Task {
  id: string;
  title: string;
  /** @nullable */
  description?: string | null;
  completed: boolean;
  /** @nullable */
  dueDate: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface TaskInput {
  /** @minLength 1 */
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  dueDate?: string | null;
}

export interface TaskUpdate {
  /** @minLength 1 */
  title?: string;
  /** @nullable */
  description?: string | null;
  completed?: boolean;
  /** @nullable */
  dueDate?: string | null;
}

export interface DashboardSummary {
  total: number;
  open: number;
  completed: number;
  dueToday: number;
  nextTasks: Task[];
}

export interface Page {
  id: string;
  /** @nullable */
  parentId: string | null;
  title: string;
  createdAt: string;
  updatedAt: string;
}

export type BlockType = typeof BlockType[keyof typeof BlockType];


export const BlockType = {
  paragraph: 'paragraph',
  heading_1: 'heading_1',
  heading_2: 'heading_2',
  heading_3: 'heading_3',
  todo: 'todo',
} as const;

export interface Block {
  id: string;
  pageId: string;
  type: BlockType;
  position: number;
  content: string;
  checked: boolean;
  createdAt: string;
  updatedAt: string;
}

export type PageDetail = Page & {
  blocks: Block[];
};

export interface PageInput {
  /** @nullable */
  parentId?: string | null;
  /** @minLength 1 */
  title?: string;
}

export interface PageUpdate {
  /** @nullable */
  parentId?: string | null;
  /** @minLength 1 */
  title?: string;
}

export type BlockInputType = typeof BlockInputType[keyof typeof BlockInputType];


export const BlockInputType = {
  paragraph: 'paragraph',
  heading_1: 'heading_1',
  heading_2: 'heading_2',
  heading_3: 'heading_3',
  todo: 'todo',
} as const;

export interface BlockInput {
  type: BlockInputType;
  position?: number;
  content?: string;
  checked?: boolean;
}

export type BlockUpdateType = typeof BlockUpdateType[keyof typeof BlockUpdateType];


export const BlockUpdateType = {
  paragraph: 'paragraph',
  heading_1: 'heading_1',
  heading_2: 'heading_2',
  heading_3: 'heading_3',
  todo: 'todo',
} as const;

export interface BlockUpdate {
  type?: BlockUpdateType;
  position?: number;
  content?: string;
  checked?: boolean;
}

export type ListTasksParams = {
status?: ListTasksStatus;
};

export type ListTasksStatus = typeof ListTasksStatus[keyof typeof ListTasksStatus];


export const ListTasksStatus = {
  all: 'all',
  open: 'open',
  completed: 'completed',
} as const;