# lib/api-zod/src/generated/types/block.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BlockType } from './blockType';

export interface Block {
  id: string;
  pageId: string;
  type: BlockType;
  position: number;
  content: string;
  checked: boolean;
  createdAt: Date;
  updatedAt: Date;
}