# lib/api-zod/src/generated/types/blockInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BlockInputType } from './blockInputType';

export interface BlockInput {
  type: BlockInputType;
  position?: number;
  content?: string;
  checked?: boolean;
}