# lib/api-zod/src/generated/types/blockUpdate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BlockUpdateType } from './blockUpdateType';

export interface BlockUpdate {
  type?: BlockUpdateType;
  position?: number;
  content?: string;
  checked?: boolean;
}