# lib/api-zod/src/generated/types/page.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Page {
  id: string;
  /** @nullable */
  parentId: string | null;
  title: string;
  createdAt: Date;
  updatedAt: Date;
}