# lib/api-zod/src/generated/types/pageDetail.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Block } from './block';
import type { Page } from './page';

export type PageDetail = Page & {
  blocks: Block[];
};