# lib/api-zod/src/generated/types/pageInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PageInput {
  /** @nullable */
  parentId?: string | null;
  /** @minLength 1 */
  title?: string;
}