# lib/api-zod/src/generated/types/pageUpdate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PageUpdate {
  /** @nullable */
  parentId?: string | null;
  /** @minLength 1 */
  title?: string;
}