# lib/api-zod/src/generated/types/task.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Task {
  id: string;
  title: string;
  /** @nullable */
  description?: string | null;
  completed: boolean;
  /** @nullable */
  dueDate: Date | null;
  createdAt: Date;
  updatedAt: Date;
}